feature: new attributes for experiment runner
